Move the old processor out of the way.
